refactor(vllm): isolate DP topology helper
